Protection of data's
This Program can be used in the beginning of any program for data hiding this helps the user to type the password with 3 guesses which if right will pass on or continue with the program. This is a efficient way to hide each individual data's or Programs for appropriate users only. This program is a simple method program which simply compares the password with the user's each guess. If it exceeds if it exceeds above the 3 Rd guess it will  break out from the code and end the program.